// src/app/core/models/appointment/appointment.model.ts
export class AppointmentModel {
    _id: string;
    number: number;
    prefix: string;
    appointmentdate: any;
    attendee: any;
    customerid: any;
    onModel: string;
    duration: number;
    charges: number;
    host: any;
    refid: any;
    timeslot: object;
    property: object;
    billid: any;
    branchid: any;
    addedby: any;
    status: string;
}